requirements.txt
hydra-core>=1.1.0
insightface==0.2.1
kornia==0.6.5
moviepy==1.0.3
onnx==1.12.0
onnxruntime==1.11.1
opencv-python==4.6.0.66
tqdm==4.64.0